Update installation.rst
openjdk-8-jdk instead of 7 since 7 has "no installation candidate" on Ubuntu16.04.

Other than that it's the same installation commands as Ubuntu15. If i remember correctly this was the only thing i changed from the Ubuntu15 commands, and now my app works normally on androids and is published. 

Just to be sure i made virtual environment with python2.7 and installed everything up to running `buildozer init`. If you want me to test it fully (that is build a new app and run it on my android to make sure no extra libs are needed) let me know.