Move Buildozer into own file.
__init__.py is responsible for:

a) Making sure clients don't need to understand the substructure of the package.
b) Determining which parts of the module should be shared with the client.
c) Due to the reg-exp based automated version detection, it also keeps track of the version of the module.

The rest of the concerns of the module have been separated into a separate file.