dependencies: rm 'virtualenv', already using "python3 -m venv"
This PR removes the `virtualenv` dependency from `setup.py`, as it looks redundant.

To be fair, I am not fully certain why setup.py was declaring `virtualenv` as needed,
I think it's a historical remnant of python 2 days.

The min required python version is atm 3.6:
https://github.com/kivy/buildozer/blob/9503f5fd153f3f08e9160799fc04b23faf6aabeb/setup.py#L16
`venv` is part of the stdlib since python 3.3, and "is now recommended for creating virtual environments" since 3.5:
https://docs.python.org/3/library/venv.html
Note that `venv` is a subset of `virtualenv`, but buildozer AFAICT is only using that subset.

Similar change in p4a:
https://github.com/kivy/python-for-android/pull/2152