Clean up inheritance of targets
1) Target class contained Android-only attribute "artifact_format". Moved that code into TargetAndroid class.

2) TargetIos had a `check_build_prepared()` method that was ineffective.

`prepare_for_build()` sets `target._build_prepared` to True.
Then `check_build_prepared()` inexplicably set it to False.
Then `build()` checks for its existence, rather than its value [this is code smell that is on my list for later], so the fact that it had been changed made no difference.

So, I removed the method, and the call to it.

2) TargetOSX had lots of methods that were identical to its base class implementations (three more, once the above steps were done). They don't need to be specified; they are inherited. I removed them. `osx.py` is now 20% smaller.